% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom Boxplot.R
\name{custom_boxplot}
\alias{custom_boxplot}
\title{Custom Box Plot Function - Made Using Highcharter Code}
\usage{
custom_boxplot(x = NULL, var = NULL, var2 = NULL, outliers = TRUE,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Custom Box Plot Function - Made Using Highcharter Code
}
