% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Custom Server.R
\name{custom_server}
\alias{custom_server}
\title{Custom Server for Shiny Dashboard}
\usage{
custom_server(input, output, session)
}
\arguments{
\item{session}{}
}
\value{

}
\description{
Custom Server for Shiny Dashboard
}
