% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny App Functions.R
\name{data_chooser}
\alias{data_chooser}
\title{Data Chooser Function - Developer Use Only}
\usage{
data_chooser(data_query)
}
\arguments{
\item{data_query}{}
}
\description{
Data Chooser Function - Developer Use Only
}
