% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny App Functions.R
\name{data_input}
\alias{data_input}
\title{Data Input - For Use in Multiple Functions}
\usage{
data_input(data = c("Education", "Household", "Health", "Employment",
  "Land/Animals", "Assets/Possessions", "Housing",
  "Vaccination & Diarrhoea", "Married Women", "Public Benefits/Services"))
}
\arguments{
\item{data}{}
}
\description{
Data Input - For Use in Multiple Functions
}
