% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny App Functions.R
\name{data_real}
\alias{data_real}
\title{Real Data Input - Default Set}
\usage{
data_real(data = list(sec_c, sec_b, sec_d, sec_e, sec_f1, sec_f2, sec_g,
  sec_h, sec_i, sec_j))
}
\arguments{
\item{data}{}
}
\description{
Real Data Input - Default Set
}
