% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny App Functions.R
\name{filter_chooser}
\alias{filter_chooser}
\title{Filter Chooser Function - Developer Use Only}
\usage{
filter_chooser(data, prov = FALSE, dist = FALSE, region = FALSE,
  element)
}
\arguments{
\item{element}{}
}
\description{
Filter Chooser Function - Developer Use Only
}
