% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend Plotting Functions.R
\name{funky_plot_chr}
\alias{funky_plot_chr}
\title{Highcharts Plot for Character Columns}
\usage{
funky_plot_chr(df, column_of_interest)
}
\arguments{
\item{column_of_interest}{}
}
\description{
Highcharts Plot for Character Columns
}
