% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend Plotting Functions.R
\name{funky_plot_dbl}
\alias{funky_plot_dbl}
\title{Highcharts Plot for Numeric Columns}
\usage{
funky_plot_dbl(df, column_of_interest, zero_omit = TRUE, out = FALSE)
}
\arguments{
\item{out}{}
}
\value{

}
\description{
Highcharts Plot for Numeric Columns
}
