% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend Filter Functions.R
\name{general_chr_filter}
\alias{general_chr_filter}
\title{Character Column Interpreter}
\usage{
general_chr_filter(data, prov = FALSE, district = FALSE,
  region = FALSE, element)
}
\arguments{
\item{element}{}
}
\value{

}
\description{
Character Column Interpreter
}
