% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend Filter Functions.R
\name{general_num_filter}
\alias{general_num_filter}
\title{Numeric Column Interpreter}
\usage{
general_num_filter(data, prov = FALSE, district = FALSE,
  region = FALSE, element)
}
\arguments{
\item{element}{}
}
\description{
Numeric Column Interpreter
}
