% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny App Functions.R
\name{info_parser}
\alias{info_parser}
\title{Info Parser Function - Developer Use Only}
\usage{
info_parser(req_info)
}
\arguments{
\item{req_info}{}
}
\description{
Info Parser Function - Developer Use Only
}
