% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny App Functions.R
\name{summary_maker}
\alias{summary_maker}
\title{Summary Maker Function - Developer Use Only}
\usage{
summary_maker(main_df, column_of_choice, zero_omit = TRUE)
}
\arguments{
\item{zero_omit}{}
}
\description{
Summary Maker Function - Developer Use Only
}
